pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nz-economic-forecasting"
version = "0.1.0"
description = "NZ retail sales forecasting using Stats NZ Aotearoa Data Explorer API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.1",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "lightgbm>=4.3",
    "prophet>=1.1",
    "mlflow>=2.11",
    "optuna>=3.6",
    "pandera>=0.18",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "pyarrow>=15.0",
    "matplotlib>=3.8",
    "statsmodels>=0.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "ipykernel>=6.29",
    "notebook>=7.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/forecasting"]
